+++ dashboard/package.json
@@ -10,7 +10,7 @@
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
     "react-router-dom": "^6.26.0",
-    "react-scripts": "5.0.1",
+    "react-scripts": "^3.0.1",
     "react-share": "^5.1.0",
     "react-svg": "^16.1.34",
     "web-vitals": "^2.1.4"
